requirements.txt
openai
anthropic
transformers
datasets
benchllama
python-dotenv
nltk
matplotlib
wandb
huggingface_hub
accelerate
torch
ollama
deepeval